apps/deploy-service/pyproject.toml
[project]
name = "build-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "docker>=7.1.0",
    "flask>=3.0.3",
    "grpcio>=1.71.0",
    "kubernetes>=31.0.0",
    "protobuf>=6.32,<7",
]

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"


[tool.ruff]
line-length = 120
extend-exclude = ["protos"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.isort]
# Use a single line after each import block.
lines-after-imports = 1

[dependency-groups]
dev = ["ruff>=0.12.10"]
